Split config finding functions to workspace/finders.py
