+++ src/tmuxp/cli/utils.py
@@ -1,12 +1,6 @@
 import logging
-import os
 import re
 import typing as t
-
-from colorama import Fore
-
-from tmuxp.types import StrPath
-from tmuxp.workspace.constants import VALID_WORKSPACE_DIR_FILE_EXTENSIONS
 
 from .. import log
 
@@ -30,162 +24,6 @@ def tmuxp_echo(
         logger.log(log.LOG_LEVELS[log_level], unstyle(message))
 
     print(message)
-
-
-def get_workspace_dir() -> str:
-    """
-    Return tmuxp configuration directory.
-
-    ``TMUXP_CONFIGDIR`` environmental variable has precedence if set. We also
-    evaluate XDG default directory from XDG_CONFIG_HOME environmental variable
-    if set or its default. Then the old default ~/.tmuxp is returned for
-    compatibility.
-
-    Returns
-    -------
-    str :
-        absolute path to tmuxp config directory
-    """
-
-    paths = []
-    if "TMUXP_CONFIGDIR" in os.environ:
-        paths.append(os.environ["TMUXP_CONFIGDIR"])
-    if "XDG_CONFIG_HOME" in os.environ:
-        paths.append(os.path.join(os.environ["XDG_CONFIG_HOME"], "tmuxp"))
-    else:
-        paths.append("~/.config/tmuxp/")
-    paths.append("~/.tmuxp")
-
-    for path in paths:
-        path = os.path.expanduser(path)
-        if os.path.isdir(path):
-            return path
-    # Return last path as default if none of the previous ones matched
-    return path
-
-
-def find_workspace_file(
-    workspace_file: StrPath,
-    workspace_dir: t.Optional[StrPath] = None,
-) -> str:
-    """
-    Return the real config path or raise an exception.
-
-    If config is directory, scan for .tmuxp.{yaml,yml,json} in directory. If
-    one or more found, it will warn and pick the first.
-
-    If config is ".", "./" or None, it will scan current directory.
-
-    If config is has no path and only a filename, e.g. "myconfig.yaml" it will
-    search config dir.
-
-    If config has no path and only a name with no extension, e.g. "myconfig",
-    it will scan for file name with yaml, yml and json. If multiple exist, it
-    will warn and pick the first.
-
-    Parameters
-    ----------
-    workspace_file : str
-        workspace file, valid examples:
-
-        - a file name, myconfig.yaml
-        - relative path, ../config.yaml or ../project
-        - a period, .
-    """
-    if not workspace_dir:
-        workspace_dir = get_workspace_dir()
-    path = os.path
-    exists, join, isabs = path.exists, path.join, path.isabs
-    dirname, normpath, splitext = path.dirname, path.normpath, path.splitext
-    cwd = os.getcwd()
-    is_name = False
-    file_error = None
-
-    workspace_file = os.path.expanduser(workspace_file)
-    # if purename, resolve to confg dir
-    if is_pure_name(workspace_file):
-        is_name = True
-    elif (
-        not isabs(workspace_file)
-        or len(dirname(workspace_file)) > 1
-        or workspace_file == "."
-        or workspace_file == ""
-        or workspace_file == "./"
-    ):  # if relative, fill in full path
-        workspace_file = normpath(join(cwd, workspace_file))
-
-    # no extension, scan
-    if path.isdir(workspace_file) or not splitext(workspace_file)[1]:
-        if is_name:
-            candidates = [
-                x
-                for x in [
-                    f"{join(workspace_dir, workspace_file)}{ext}"
-                    for ext in VALID_WORKSPACE_DIR_FILE_EXTENSIONS
-                ]
-                if exists(x)
-            ]
-            if not len(candidates):
-                file_error = (
-                    "workspace-file not found in workspace dir (yaml/yml/json) %s "
-                    "for name" % (workspace_dir)
-                )
-        else:
-            candidates = [
-                x
-                for x in [
-                    join(workspace_file, ext)
-                    for ext in [".tmuxp.yaml", ".tmuxp.yml", ".tmuxp.json"]
-                ]
-                if exists(x)
-            ]
-
-            if len(candidates) > 1:
-                tmuxp_echo(
-                    Fore.RED
-                    + "Multiple .tmuxp.{yml,yaml,json} workspace_files in %s"
-                    % dirname(workspace_file)
-                    + Fore.RESET
-                )
-                tmuxp_echo(
-                    "This is undefined behavior, use only one. "
-                    "Use file names e.g. myproject.json, coolproject.yaml. "
-                    "You can load them by filename."
-                )
-            elif not len(candidates):
-                file_error = "No tmuxp files found in directory"
-        if len(candidates):
-            workspace_file = candidates[0]
-    elif not exists(workspace_file):
-        file_error = "file not found"
-
-    if file_error:
-        raise FileNotFoundError(file_error, workspace_file)
-
-    return workspace_file
-
-
-def is_pure_name(path: str) -> bool:
-    """
-    Return True if path is a name and not a file path.
-
-    Parameters
-    ----------
-    path : str
-        Path (can be absolute, relative, etc.)
-
-    Returns
-    -------
-    bool
-        True if path is a name of config in config dir, not file path.
-    """
-    return (
-        not os.path.isabs(path)
-        and len(os.path.dirname(path)) == 0
-        and not os.path.splitext(path)[1]
-        and path != "."
-        and path != ""
-    )
 
 
 def prompt(
