+++ src/tmuxp/workspace/config.py
@@ -44,76 +44,6 @@ def validate_schema(workspace_dict):
     return True
 
 
-def is_workspace_file(filename, extensions=[".yml", ".yaml", ".json"]):
-    """
-    Return True if file has a valid workspace file type.
-
-    Parameters
-    ----------
-    filename : str
-        filename to check (e.g. ``mysession.json``).
-    extensions : str or list
-        filetypes to check (e.g. ``['.yaml', '.json']``).
-
-    Returns
-    -------
-    bool
-    """
-    extensions = [extensions] if isinstance(extensions, str) else extensions
-    return any(filename.endswith(e) for e in extensions)
-
-
-def in_dir(
-    config_dir=os.path.expanduser("~/.tmuxp"), extensions=[".yml", ".yaml", ".json"]
-):
-    """
-    Return a list of configs in ``config_dir``.
-
-    Parameters
-    ----------
-    config_dir : str
-        directory to search
-    extensions : list
-        filetypes to check (e.g. ``['.yaml', '.json']``).
-
-    Returns
-    -------
-    list
-    """
-    configs = []
-
-    for filename in os.listdir(config_dir):
-        if is_workspace_file(filename, extensions) and not filename.startswith("."):
-            configs.append(filename)
-
-    return configs
-
-
-def in_cwd():
-    """
-    Return list of configs in current working directory.
-
-    If filename is ``.tmuxp.py``, ``.tmuxp.json``, ``.tmuxp.yaml``.
-
-    Returns
-    -------
-    list
-        configs in current working directory
-
-    Examples
-    --------
-    >>> sorted(in_cwd())
-    ['.tmuxp.json', '.tmuxp.yaml']
-    """
-    configs = []
-
-    for filename in os.listdir(os.getcwd()):
-        if filename.startswith(".tmuxp") and is_workspace_file(filename):
-            configs.append(filename)
-
-    return configs
-
-
 def expandshell(_path):
     """
     Return expanded path based on user's ``$HOME`` and ``env``.
